Simplify the dev footer display
